@@ -5,60 +5,63 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "start-vm": "ng serve --env=dev.vm",
+    "start-vm": "ng serve --configuration=dev.vm",
     "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "build-prod": "ng build --prod --output-path=build-prod --env=prod",
-    "build-dev": "ng build --dev --output-path=build-dev --env=dev",
+    "build-prod": "ng build --prod --output-path=build-prod --configuration=production",
+    "build-dev": "ng build --dev --output-path=build-dev --configuration=dev",
     "compodoc": "compodoc -p src/tsconfig.app.json -w -s --includes ./documentation-extended -a ./documentation-extended/assets --name QMDocs",
     "compodoc-build": "compodoc -p src/tsconfig.app.json --includes ./documentation-extended -a ./documentation-extended/assets --name QMDocs"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.5",
-    "@angular/cdk": "^5.2.1",
-    "@angular/common": "^5.2.3",
-    "@angular/compiler": "^5.2.3",
-    "@angular/core": "^5.2.3",
-    "@angular/forms": "^5.2.3",
-    "@angular/http": "^5.2.3",
-    "@angular/material": "^5.2.1",
-    "@angular/platform-browser": "^5.2.3",
-    "@angular/platform-browser-dynamic": "^5.2.3",
-    "@angular/router": "^5.2.3",
-    "angularfire2": "^5.0.0-rc.5-next",
-    "bootstrap": "^4.1.0",
-    "core-js": "^2.4.1",
-    "firebase": "^4.9.1",
+    "@angular/animations": "^6.1.0",
+    "@angular/cdk": "^6.4.2",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/material": "^6.4.2",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
+    "angularfire2": "^5.0.0-rc.11",
+    "bootstrap": "^4.1.3",
+    "core-js": "^2.5.7",
+    "firebase": "^5.3.0",
     "font-awesome": "^4.7.0",
+    "hoek": "^5.0.3",
     "jquery": "^3.3.1",
-    "moment": "^2.22.0",
-    "ngx-bootstrap": "^2.0.3",
-    "ngx-clipboard": "^10.0.0",
+    "moment": "^2.22.2",
+    "ngx-bootstrap": "^3.0.1",
+    "ngx-clipboard": "^11.1.1",
     "node-sass": "^4.9.2",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.0.0-rc.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.3",
-    "@angular/language-service": "^5.2.3",
-    "@types/jasmine": "~2.8.3",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "^6.1.1",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~9.6.2",
-    "codelyzer": "^4.0.1",
+    "@types/node": "~10.5.5",
+    "codelyzer": "^4.4.2",
     "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^2.0.4",
+    "karma": "^2.0.5",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^1.0.0",
-    "protractor": "^5.3.2",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "^2.6.2"
+    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^1.2.0",
+    "protractor": "^5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "^2.7.2"
   }
 }
