@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {AngularFireDatabase} from 'angularfire2/database';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {User} from '../../shared/model/user';
 import {UserService} from '../../core/user.service';
 import {MatSnackBar} from "@angular/material";
