@@ -1,13 +1,11 @@
+import {catchError, map, switchMap, tap} from 'rxjs/operators';
+import {BehaviorSubject, Observable, of, throwError} from 'rxjs';
 import {Injectable} from '@angular/core';
 import {AngularFireDatabase} from 'angularfire2/database';
-import {Observable} from 'rxjs/Observable';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
-import {catchError, tap} from 'rxjs/operators';
 import {LogService} from './log.service';
 import {environment} from "../../environments/environment";
-import {BehaviorSubject} from "rxjs/BehaviorSubject";
 import {User} from "../shared/model/user";
-import 'rxjs/add/observable/throw';
 import {Helper} from "../shared/helper/helper";
 import {ActiondId} from "../shared/model/actionEntryLog";
 import {Detail} from "../shared/model/detail";
@@ -20,7 +18,7 @@ export class UserService {
     })
   };
   /* ERROR MESSAGES */
-  USER_NOT_FOUND: string = "USER NOT FOUND";
+  USER_NOT_FOUND = "USER NOT FOUND";
   private userapi: string = environment.api + '/users';
   private qmapi: string = environment.api + '/users/qm';
   private incidentapi: string = environment.api + '/incidents';
@@ -29,54 +27,61 @@ export class UserService {
 
   constructor(public db: AngularFireDatabase,
               public http: HttpClient,
-              public logService: LogService,
-  ) {
+              public logService: LogService,) {
   }
 
   getUsers(): Observable<User[]> {
-    return this.http.get(this.userapi)
-      .map((resp: any) => {
-          // verify response
-          if (resp.code === 200) {
-            return resp.data.map(el =>
-              // populate the User Model
-              new User(el.user_id, el.first_name, el.last_name, el.is_available, el.current_q_days, el.incident_threshold, el.usage_percent, el.incident_counts, el.supported_products)
-            )
-          } else {
-            throw new Error(resp.code)
-          }
-      }).pipe(
-        catchError((e) => Helper.handleError(e, "Failed to get users"))
-      )
+    return this.http.get(this.userapi).pipe(
+      map((resp: any) => {
+        // verify response
+        if (resp.code === 200) {
+          return resp.data.map(el =>
+            // populate the User Model
+            new User(el.user_id, el.first_name, el.last_name, el.is_available,
+              el.current_q_days, el.incident_threshold, el.usage_percent,
+              el.incident_counts, el.supported_products)
+          );
+        } else {
+          throw new Error(resp.code);
+        }
+      })).pipe(
+      catchError((e) => Helper.handleError(e, "Failed to get users"))
+    );
   }
 
-  getUserByNumber(iNumber: string): Observable<User> {
-    if (!iNumber) throw new Error("Empty Argument");
+  getUserByNumber(iNumber: string): Observable<any> {
+    if (!iNumber) {
+      throwError("Empty Argument");
+    }
     const url = `${this.userapi}/${iNumber}`;
-    return this.http.get(url).map((resp: any) => {
-      if (resp.code === 200) {
-        return new User(resp.data.user_id, resp.data.first_name, resp.data.last_name, resp.data.is_available, resp.data.current_q_days, resp.data.incident_threshold, resp.usage_perecent, resp.data.incident_counts, resp.data.supported_products)
-      } else {
-        throw new Error("User not found")
-      }
-    })
-      .pipe(
-        catchError((e) => Helper.handleError(e, "Failed to get user"))
-      )
+    return this.http.get(url).pipe(
+      map((resp: any) => {
+        if (resp.code === 200) {
+          return new User(resp.data.user_id, resp.data.first_name, resp.data.last_name,
+            resp.data.is_available, resp.data.current_q_days, resp.data.incident_threshold,
+            resp.usage_perecent, resp.data.incident_counts, resp.data.supported_products);
+        } else {
+          throwError("User not found");
+        }
+      }),
+      catchError((e) => Helper.handleError(e, "Failed to get user"))
+    )
   }
 
   addUser(firstName: string, iNumber: string): Observable<User> {
-    let body = {
+    const body = {
       "user_id": iNumber,
       "first_name": firstName,
       "last_name": ""
     };
-    return this.http.post(this.userapi, body, this.httpOptions).switchMap((resp:any) => {
-      if (resp.code == 201) return this.getUserByNumber(iNumber);
+    return this.http.post(this.userapi, body, this.httpOptions).pipe(switchMap((resp: any) => {
+      if (resp.code == 201) {
+        return this.getUserByNumber(iNumber);
+      }
       throw Error();
-    }).pipe(
+    })).pipe(
       catchError(e => Helper.handleError(e, "Failed to Add User"))
-    )
+    );
   }
 
   updateUserMeta(user: User) {
@@ -84,11 +89,11 @@ export class UserService {
     return this.http.put(`${this.userapi}/${user.iNumber}`, body, this.httpOptions)
       .pipe(
         catchError(e => Helper.handleError(e, "Update User Meta Failed"))
-      )
+      );
   }
 
   updateAvailability(user: User, bool: boolean): Observable<any> {
-    let body = this.buildBodyFromUserObject(user);
+    const body = this.buildBodyFromUserObject(user);
     body.is_available = bool;
 
     return this.http.put(this.userapi + '/' + user.iNumber, body, this.httpOptions)
@@ -96,43 +101,43 @@ export class UserService {
         tap(() => this.logService.addLog(user, ActiondId.AVAILABILITY_CHANGED,
           new Detail(user.getStatus(), user.isAvailable ? "BUSY" : "AVAILABLE", ""))),
         catchError(e => Helper.handleError(e, "Update Availability Failed"))
-      )
+      );
   }
 
   deleteUser(iNumber: string) {
     return this.http.delete(`${this.userapi}/${iNumber}`)
       .pipe(
         catchError(e => Helper.handleError(e, "Failed to delete user"))
-      )
+      );
   }
 
   updateSupport(user: User, productShortName: string, bool: boolean) {
     const body = {
       "supported": bool
     };
-    return this.http.put(`${this.userapi}/${user.iNumber}/${productShortName}`, body, this.httpOptions).map((res: any) => {
+    return this.http.put(`${this.userapi}/${user.iNumber}/${productShortName}`, body, this.httpOptions).pipe(map((res: any) => {
       if (res.code === 200) {
         return res;
       } else {
         throw new Error();
       }
-    })
+    }))
       .pipe(
         tap(() => this.logService.addLog(user, ActiondId.PRDOUCT_SUPPORT_CHANGED,
           new Detail().addCustomDetail(user.hasRole(productShortName) ? "Removed " + productShortName : "Added " + productShortName))),
-        catchError(e => Helper.handleError(e, "Failed to update support")))
+        catchError(e => Helper.handleError(e, "Failed to update support")));
   }
 
   addIncident(user: User, productShortName: string): Observable<any> {
-    let body = {
+    const body = {
       "product_short_name": productShortName,
       "user_id": user.iNumber
     };
     return this.http.post(this.incidentapi, body, this.httpOptions)
       .pipe(
-        tap(()=> this.logService.addLog(user,ActiondId.INCIDENT_ASSIGNED,
-          new Detail(user.incidentCounts[productShortName],user.incidentCounts[productShortName]+1, productShortName))),
-        catchError(e => Helper.handleError(e, "Add Incident Failed")))
+        tap(() => this.logService.addLog(user, ActiondId.INCIDENT_ASSIGNED,
+          new Detail(user.incidentCounts[productShortName], user.incidentCounts[productShortName] + 1, productShortName))),
+        catchError(e => Helper.handleError(e, "Add Incident Failed")));
   }
 
   removeIncident(user: User, productShortName: string): Observable<any> {
@@ -142,14 +147,14 @@ export class UserService {
     };
     return this.http.delete(this.incidentapi + `/${user.iNumber}/${productShortName}`, this.httpOptions)
       .pipe(
-        tap(()=> this.logService.addLog(user,ActiondId.INCIDENT_UNASSIGNED,
-          new Detail(user.incidentCounts[productShortName],user.incidentCounts[productShortName]-1, productShortName))),
-        catchError(e => Helper.handleError(e, "Remove Incident Failed")))
+        tap(() => this.logService.addLog(user, ActiondId.INCIDENT_UNASSIGNED,
+          new Detail(user.incidentCounts[productShortName], user.incidentCounts[productShortName] - 1, productShortName))),
+        catchError(e => Helper.handleError(e, "Remove Incident Failed")));
   }
 
   // TODO Refactor out, redundant and sub-clone of #Update Queue Days
   restQueueDays(user: User) {
-    return Observable.of(5)
+    return of(5);
     // let tmp = new User(user);
     // tmp.currentQDays = 0;
     // return this.userSetService.resetRCC(tmp)
@@ -157,25 +162,33 @@ export class UserService {
   }
 
   updateQueueDays(user, amount) {
-    let requestUser = User.copy(user);
+    const requestUser = User.copy(user);
     requestUser.currentQDays = amount;
-    return this.updateUserMeta(requestUser).map(() => amount)
+    return this.updateUserMeta(requestUser).pipe(map(() => amount))
       .pipe(
         tap(() => this.logService.addLog(user, ActiondId.QUEUE_DAYS_CHANGED,
           new Detail(user.currentQDays, amount, "Queue Days"))),
         catchError(e => Helper.handleError(e, "Update Queue Days Failed")));
   }
 
+
   getQM(): Observable<User> {
-    return this.http.get(this.qmapi).map((resp: any) => {
-      if (resp.code !== 200) throw new Error("error");
-      return new User(resp.data.user_id, resp.data.first_name, resp.data.last_name, resp.data.is_available, resp.data.current_q_days, resp.data.incident_threshold, resp.usage_percent, resp.data.incident_counts, resp.data.supported_products)
-    }).pipe(catchError(e => Helper.handleError(e, "Failed to get QM")))
+    return this.http.get<User>(this.qmapi).pipe(
+      map((r: any) => {
+        if (r.code !== 200) {
+          throwError("error")
+        }
+        return new User(r.data.user_id, r.data.first_name, r.data.last_name,
+          r.data.is_available, r.data.current_q_days, r.data.incident_threshold,
+          r.usage_percent, r.data.incident_counts, r.data.supported_products)
+      }),
+      catchError((e: any) => Helper.handleError(e, "Failed to get QM"))
+    )
   }
 
   setQM(iNumber: string) {
     let user;
-    return this.getUserByNumber(iNumber).switchMap(
+    return this.getUserByNumber(iNumber).pipe(switchMap(
       (u: User) => {
         user = u;
         // noinspection SpellCheckingInspection
@@ -184,14 +197,14 @@ export class UserService {
         };
         return this.http.put(this.qmapi, body, this.httpOptions);
       }
-    ).pipe(
+    )).pipe(
       tap(() => this.logService.addLog(user, ActiondId.QM_CHANGED,
         new Detail().addCustomDetail("QM Changed to " + user.name()))),
-      catchError(e => Helper.handleError(e, "Failed to set QM")))
+      catchError(e => Helper.handleError(e, "Failed to set QM")));
   }
 
-  getUserIncidents(iNumber:string){
-    return this.http.get(this.userapi + `/${iNumber}/incidents`).pipe(catchError(e => Helper.handleError(e, "Failed to get Incidents")))
+  getUserIncidents(iNumber: string) {
+    return this.http.get(this.userapi + `/${iNumber}/incidents`).pipe(catchError(e => Helper.handleError(e, "Failed to get Incidents")));
   }
 
   private buildBodyFromUserObject(user: User) {
@@ -203,7 +216,7 @@ export class UserService {
       "usage_percent": user.usagePercent,
       "current_q_days": user.currentQDays,
       "incident_threshold": user.iThreshold
-    }
+    };
   }
 
 }
