@@ -1,6 +1,9 @@
+
+import {of as observableOf, Observable} from 'rxjs';
+
+import {map} from 'rxjs/operators';
 import {Component, OnInit} from '@angular/core';
 import {LogService} from '../../../core/log.service';
-import {Observable} from 'rxjs/Observable';
 import {ActionEntryLog} from "../../model/actionEntryLog";
 
 @Component({
@@ -20,9 +23,9 @@ export class ActivityLogComponent implements OnInit {
   }
 
   ngOnInit() {
-    this.activityLog$ = Observable.of([]);
+    this.activityLog$ = observableOf([]);
 
-    this.logService.getLogs().map(arr => arr.slice(0, this.NUM_OF_RESULTS)).subscribe((res: any) => {
+    this.logService.getLogs().pipe(map(arr => arr.slice(0, this.NUM_OF_RESULTS))).subscribe((res: any) => {
       this.actionActiveLogList = res;
     })
   }
