@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {BsModalRef} from 'ngx-bootstrap/modal/bs-modal-ref.service';
-import {Subject} from "rxjs/Subject";
+import {Subject} from "rxjs";
 
 
 @Component({
