@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {Subject} from "rxjs/Subject";
+import {Subject} from "rxjs";
 import {BsModalRef} from "ngx-bootstrap/modal/bs-modal-ref.service";
 
 @Component({
